+++ CHANGELOG.md
@@ -3,6 +3,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [Unreleased]
+
+### Changes
+
+- Upgraded to timely dataflow and differential dataflow dependencies to v0.12.
+
 ## [0.39.0] - April 11, 2021
 
 ### D3log